+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Changelog
 - [FIR, Anvil Interop] Fix rank-based binding replacements getting dropped for multi-contribution classes in root graphs when contributions are from external modules.
 - [IR, Anvil Interop] Fix rank-based binding replacements getting dropped for multi-contribution classes in graph extensions when contributions are from external modules.
 - [FIR] Named annotation arguments in different order from declared parameters getting silently skipped.
+- [IR] Prohibit duplicate map keys.
 
 0.9.2
 -----